Fix split package in mapper-extras module
The mapper-extras module reuses server package names for
its classes. This commit moves these classes into a new extras
package just for this module.
