+++ modules/mapper-extras/build.gradle
@@ -11,27 +11,11 @@ apply plugin: 'elasticsearch.internal-cluster-test'
 
 esplugin {
   description 'Adds advanced field mappers'
-  classname 'org.elasticsearch.index.mapper.MapperExtrasPlugin'
+  classname 'org.elasticsearch.index.mapper.extras.MapperExtrasPlugin'
 }
 
 restResources {
   restApi {
     include '_common', 'cluster', 'field_caps', 'nodes', 'indices', 'index', 'search', 'get'
   }
-}
-
-tasks.named('splitPackagesAudit').configure {
-  // index is owned by server, move these to mapper specific packages
-  ignoreClasses 'org.elasticsearch.index.query.RankFeatureQueryBuilder',
-    'org.elasticsearch.index.query.RankFeatureQueryBuilders',
-    'org.elasticsearch.index.query.SourceConfirmedTextQuery',
-    'org.elasticsearch.index.query.SourceIntervalsSource',
-    'org.elasticsearch.index.mapper.MapperExtrasPlugin',
-    'org.elasticsearch.index.mapper.MatchOnlyTextFieldMapper',
-    'org.elasticsearch.index.mapper.RankFeatureFieldMapper',
-    'org.elasticsearch.index.mapper.RankFeatureMetaFieldMapper',
-    'org.elasticsearch.index.mapper.RankFeaturesFieldMapper',
-    'org.elasticsearch.index.mapper.ScaledFloatFieldMapper',
-    'org.elasticsearch.index.mapper.SearchAsYouTypeFieldMapper',
-    'org.elasticsearch.index.mapper.TokenCountFieldMapper'
 }
