+++ modules/mapper-extras/src/test/resources/org/elasticsearch/index/mapper/extras/filebeat-6.0.template.json
