+++ modules/mapper-extras/src/test/resources/org/elasticsearch/index/mapper/extras/metricbeat-6.0.template.json
