+++ modules/mapper-extras/src/test/resources/org/elasticsearch/index/mapper/extras/packetbeat-6.0.template.json
